Added update mode switching.
Created the `thorough` and `onwrite` profiles.

These profiles govern the hooks -- autocommands and mappings --
that jslint.vim will respond to with update events.

It is possible to switch a buffer's update profile
with the new `:JSLintUpdateMode` command.

The `thorough` profile, enabled by default,
sets up buffers as they have been set up.

It may be useful to people with slow machines, large JS scripts,
and/or slow JS interpreters
to have the lint checked less often, so I added the `onwrite` profile.

Use of the `onwrite` profile alleviates issues I was having
where the quickfix entries would not work properly:

About half of the time, quickfix entry selection
-- via `:cc`, pressing <Enter> on a qf entry, etc. --
would go to the wrong line.

The cursor would end up
on the line of another entry in the quickfix list.

Disabling the BufEnter autocommand fixes this.